[skip changelog] Separate block validation related code
